process: isolate state marshaling
